container_image_code/requirements.txt
boto3
Pillow
tensorflow-cpu